crud완성

+++ src/main/java/com/simple/spring/dao/AnalectsRepository.java
@@ -9,16 +9,16 @@ import org.springframework.stereotype.Repository;
 import com.simple.spring.model.Analects;
 
 @Repository
-public interface AnalectsRepository extends CrudRepository<Analects, Integer> {
+public interface AnalectsRepository extends CrudRepository<Analects, Long> {
 		
 	@Override // Create, Update
 	Analects save(Analects anal);
 
     @Override // Read
-    Optional<Analects> findById(Integer id);
+    Optional<Analects> findById(Long id);
     @Override // Read All
     List<Analects> findAll();
 
     @Override // Delete
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
